Revert "added population living difference queries (#77)"
This reverts commit 922833e9df24d95c4c88dc599d74c748ef749b2e.

+++ src/main/java/com/napier/sem/CapitalCityReportView.java
@@ -19,13 +19,13 @@ public class CapitalCityReportView extends ReportView {
     /**
      * City Population
      */
-    public long Population;
+    public int Population;
 
     public CapitalCityReportView(ResultSet results) {
         try {
             Name = results.getString("Name");
             Country = results.getString("Country");
-            Population = results.getLong("Population");
+            Population = results.getInt("Population");
         } catch (Exception e) {
             System.out.println(e.getMessage());
             System.out.println("Failed to fetch country report view");
